Added functionality + switched api resources
Has all the desired functionality. Switched to using the API resources that YR is linking on its website i.e. forecast and symbol urls. This guarantees that the searches will work however this is not all search options that YR themselves have. They don't use all cities that are on geonames.org and they don't use as few cities as in the files that they share on their website. Very annoying.  Would like to have a complete list of all cities that are eligible to get forecast data for.